+++ lib/riaction/crud_event_callback.rb
@@ -8,19 +8,19 @@ module Riaction
     end
 
     def after_create(record)
-      if record.riaction_log_event?(@event_name)
+      if record.riaction_log_event?(@event_name) && !record.class.riactionless?
         Resque.enqueue(::Riaction::EventPerformer, @event_name, record.class.base_class.to_s, record.id)
       end
     end                                 
 
     def after_update(record)            
-      if record.riaction_log_event?(@event_name)
+      if record.riaction_log_event?(@event_name) && !record.class.riactionless?
         Resque.enqueue(::Riaction::EventPerformer, @event_name, record.class.base_class.to_s, record.id)
       end
     end
 
     def after_destroy(record)
-      if record.riaction_log_event?(@event_name)
+      if record.riaction_log_event?(@event_name) && !record.class.riactionless?
         Resque.enqueue(::Riaction::EventPerformer, @event_name, record.class.base_class.to_s, record.id)
       end
     end