+++ lib/riaction/profile_creation_callback.rb
@@ -4,7 +4,7 @@ require 'riaction/profile_creator'
 module Riaction
   class ProfileCreationCallback
     def after_create(record)
-      Resque.enqueue(::Riaction::ProfileCreator, record.class.base_class.to_s, record.id)
+      Resque.enqueue(::Riaction::ProfileCreator, record.class.base_class.to_s, record.id) unless record.class.riactionless?
     end                                 
   end
 end